Migrate to wayland
This replaces Openbox with the KWin compositor, which runs standalone
without KDE. More lightweight compositors are not usable due to a bug in
qemu which breaks mouse cursor rendering:

* https://gitlab.com/qemu-project/qemu/-/issues/2315

Weston, Mutter and KWin are the only compositors which work around this
qemu bug. Weston is the lightest of them, but can't switch output modes
at runtime. The latter one is required to make VM windows resizable:

* https://gitlab.freedesktop.org/wayland/weston/-/issues/339
* https://gitlab.freedesktop.org/wayland/weston/-/issues/341

Mutter seems to have hidden hard dependencies on GNOME, which breaks
certain features like keyboard layout switching, IBus, etc.

This leaves KWin as the only viable option.